% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_scales_discrete.R
\name{scale_colour_paletteer_d}
\alias{scale_colour_paletteer_d}
\alias{scale_color_paletteer_d}
\alias{scale_fill_paletteer_d}
\title{Discrete scales to use for ggplot2}
\usage{
scale_colour_paletteer_d(package, palette, dynamic = FALSE, ...)

scale_color_paletteer_d(package, palette, dynamic = FALSE, ...)

scale_fill_paletteer_d(package, palette, dynamic = FALSE, ...)
}
\arguments{
\item{package}{Name of package from which the palette is desired.}

\item{palette}{Name of palette.}

\item{dynamic}{toggles between the discrete palettes and the dynamic
palettes. Defaults to FALSE which indicates discrete palettes.}

\item{...}{additional arguments to pass to discrete_scale}
}
\description{
These functions provide the option to use the discrete and dynamic palettes
along with the \code{ggplot2} package. It goes without saying that it requires
\code{ggplot2} to work.
}
