% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_scales.R
\name{ggplot2-scales}
\alias{ggplot2-scales}
\alias{scale_colour_paletteer_c}
\alias{scale_color_paletteer_c}
\alias{scale_fill_paletteer_c}
\title{Scales to use for ggplot2}
\usage{
scale_colour_paletteer_c(package, palette, ...)

scale_color_paletteer_c(package, palette, ...)

scale_fill_paletteer_c(package, palette, ...)
}
\arguments{
\item{package}{Name of package from which the palette is desired.}

\item{palette}{Name of palette.}

\item{...}{Arguments to pass on to \code{ggplot2::scale_colour_gradientn()} or
\code{ggplot2::scale_fill_gradientn()}}
}
\value{
A \code{ScaleContinuous} object that can be added to a \code{ggplot} object
}
\description{
These functions provide the option to use continious palettes along with the
\code{ggplot2} package. It goes without saying that it requires \code{ggplot2} to work.
}
