% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paletteer_c_wrappers.R
\name{paleteer-c-wrapper}
\alias{paleteer-c-wrapper}
\alias{paletteer_c_ggthemes}
\alias{paletteer_c_grDevices}
\alias{paletteer_c_oompaBase}
\alias{paletteer_c_palr}
\alias{paletteer_c_pals}
\alias{paletteer_c_scico}
\alias{paletteer_c_viridis}
\alias{paletteer_c_harrypotter}
\alias{paletteer_c_gameofthrones}
\title{Wrappers around continuous palette functions by package}
\usage{
paletteer_c_ggthemes(name, n)

paletteer_c_grDevices(name, n)

paletteer_c_oompaBase(name, n)

paletteer_c_palr(name, n)

paletteer_c_pals(name, n)

paletteer_c_scico(name, n)

paletteer_c_viridis(name, n)

paletteer_c_harrypotter(name, n)

paletteer_c_gameofthrones(name, n)
}
\arguments{
\item{name}{Character, name of palette.}

\item{n}{Integer, number of colors.}
}
\value{
Vector of color values from specified palette.
}
\description{
These functions provide a function wrapper for each package such that a
palette can be generated using only the name of the desired palette and the
length.
}
