% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paletteer-package.R
\docType{package}
\name{paletteer-package}
\alias{paletteer}
\alias{paletteer-package}
\title{paletteer: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EmilHvitfeldt/paletteer}
  \item Report bugs at \url{https://github.com/EmilHvitfeldt/paletteer/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
