% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{palettes_c_names}
\alias{palettes_c_names}
\title{Names of all continuous palettes}
\format{A data.frame of 175 observations with 2 variables:
\describe{
\item{package}{character, name of package}
\item{palette}{character, name of palette}
}}
\usage{
palettes_c_names
}
\description{
A data.frame of the names and lengths of the continuous palettes included in
this package.
}
\details{
Contains palettes names from the following packages: ggthemes, grDevices,
jcolors, oompaBase, palr, pals, scico and viridis.
}
\keyword{datasets}
