% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{palettes_d_names}
\alias{palettes_d_names}
\title{Names of all fixed discrete palettes}
\format{A data.frame of 785 observations with 3 variables:
\describe{
\item{package}{character, name of package}
\item{palette}{character, name of palette}
\item{length}{integer, number of colors in palette}
}}
\usage{
palettes_d_names
}
\description{
A data.frame of the names and lengths of the discrete palettes of fixed
lengths included in this package.
}
\details{
Contains palettes names from the following packages: awtools, dichromat,
dutchmasters, ggsci, ggpomological, ggthemes, ghibli, grDevices, jcolors,
LaCroixColoR, NineteenEightyR, nord, ochRe, pals, Polychrome, quickpalette,
rcartocolor, RColorBrewer, Redmonder, RSkittleBrewer, wesanderson and yarrr.

Full list of palette can be found here \code{\link{palettes_d}}.
}
\keyword{datasets}
