% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{palettes_dynamic}
\alias{palettes_dynamic}
\title{Complete list of dynamic palettes}
\format{A list of 3 lists.}
\usage{
palettes_dynamic
}
\description{
A list of all the dynamic palettes included in this package. Structured as
a list of lists with the first level being each package and the second level
being the palettes avaliable in that package, and the third level being the
palettes for various lengths.
}
\details{
Contains palettes from the following packages: cartography and ggthemes.

Full list of palette names can be found here \code{\link{palettes_dynamic_names}}.
}
\keyword{datasets}
