% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paletteer_c_wrappers.R
\docType{data}
\name{wrapper_c}
\alias{wrapper_c}
\title{List of all wrapper functions used in paletteer_c}
\format{
Named list of all wrapper functions
}
\usage{
wrapper_c
}
\description{
List of all wrapper functions used in paletteer_c
}
\keyword{datasets}
