% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paletteer-package.R, R/paletteer.R
\docType{package}
\name{paletteer-package}
\alias{paletteer-package}
\title{paletteer: Comprehensive Collection of Color Palettes}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The choices of color palettes in R can be quite overwhelming with palettes spread over many packages with many different API's. This packages aims to collect all color palettes across the R ecosystem under the same package with a streamlined API.

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The choices of color palettes in R can be quite overwhelming with palettes spread over many packages with many different API's. This packages aims to collect all color palettes across the R ecosystem under the same package with a streamlined API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EmilHvitfeldt/paletteer}
  \item Report bugs at \url{https://github.com/EmilHvitfeldt/paletteer/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/EmilHvitfeldt/paletteer}
  \item Report bugs at \url{https://github.com/EmilHvitfeldt/paletteer/issues}
}

}
\keyword{internal}
