% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paletteer_c.R
\name{paletteer_c}
\alias{paletteer_c}
\title{Get continuous palette by package and name}
\usage{
paletteer_c(package, palette, n, scale = FALSE)
}
\arguments{
\item{package}{Name of package from which the palette is desired.}

\item{palette}{Name of palette.}

\item{n}{Number of colors desired. If omitted, returns complete palette.}

\item{scale}{Toggles quotation, defaults to FALSE.}
}
\value{
A vector of colors.
}
\description{
Get continuous palette by package and name
}
\examples{
paletteer_c(jcolors, pal10, 10)
paletteer_c(scico, berlin, 100)
}
