% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paletteer_c.R
\name{paletteer_c}
\alias{paletteer_c}
\title{Get continuous palette by package and name}
\usage{
paletteer_c(package, palette, n, direction = 1, scale = FALSE)
}
\arguments{
\item{package}{Name of package from which the palette is desired as string
or symbol.}

\item{palette}{Name of palette as string or symbol.}

\item{n}{Number of colors desired. If omitted, returns complete palette.}

\item{direction}{Either \code{1} or \code{-1}. If \code{-1} the palette will be reversed.}

\item{scale}{Toggles quotation, defaults to FALSE.}
}
\value{
A vector of colors.
}
\description{
Available package/palette combinations are available in the data.frame
\code{\link[paletteer]{palettes_c_names}}. Both \code{package} and \code{palette} can
be supplied as symbols or strings.
}
\examples{
paletteer_c(jcolors, pal10, 10)
paletteer_c(scico, berlin, 100)
}
