% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paletteer_dynamic.R
\name{paletteer_dynamic}
\alias{paletteer_dynamic}
\title{Get dynamic palette by package and name}
\usage{
paletteer_dynamic(palette, n, direction = 1)
}
\arguments{
\item{palette}{Name of palette as string or symbol. Must be on the form
packagename::palettename.}

\item{n}{Number of colors desired. If omitted, returns complete palette.}

\item{direction}{Either \code{1} or \code{-1}. If \code{-1} the palette will be reversed.}
}
\value{
A vector of colors.
}
\description{
Available package/palette combinations are available in the data.frame
\code{\link[paletteer]{palettes_dynamic_names}}. Both \code{package} and \code{palette} can
be supplied as symbols or strings.
}
\examples{
paletteer_dynamic("ggthemes_solarized::green", 8)
paletteer_dynamic("cartography::sand.pal", 20)
paletteer_dynamic(`cartography::sand.pal`, 20)
}
