% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paletteer_dynamic.R
\name{paletteer_dynamic}
\alias{paletteer_dynamic}
\title{Get dynamic palette by package and name}
\usage{
paletteer_dynamic(package, name, n)
}
\arguments{
\item{package}{Name of package from which the palette is desired.}

\item{name}{Name of palette.}

\item{n}{Number of colors desired. If omitted, returns complete palette.}
}
\value{
A vector of colors.
}
\description{
Get dynamic palette by package and name
}
\examples{
paletteer_dynamic("solarized", "green", 8)
paletteer_dynamic("cartography", "sand.pal", 20)
}
