% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{palettes_d}
\alias{palettes_d}
\title{Complete list of fixed discrete palettes}
\format{
A list of 53 lists.
}
\usage{
palettes_d
}
\description{
A list of all the discrete palettes of fixed lengths included in this
package. Structured as a list of lists with the first level being each
package and the second level being the palettes available in that package.
}
\details{
Contains palettes from the following packages:
awtools, basetheme, beyonce, calecopal, colorBlindness, colorblindr, colRoz, dichromat, DresdenColor, dutchmasters, fishualize, futurevisions, ggpomological, ggsci, ggthemes, ggthemr, ghibli, grDevices, IslamicArt, jcolors, khroma, LaCroixColoR, lisa, Manu, MapPalettes, miscpalettes, nationalparkcolors, nbapalettes, NineteenEightyR, nord, ochRe, palettesForR, palettetown, pals, PNWColors, Polychrome, rcartocolor, RColorBrewer, Redmonder, rockthemes, RSkittleBrewer, rtist, soilpalettes, suffrager, tidyquant, trekcolors, tvthemes, unikn, vapeplot, vapoRwave, werpals, wesanderson, yarrr.

data.frame of palette names can be found here \code{\link{palettes_d_names}}.
}
\keyword{datasets}
