% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{palettes_d}
\alias{palettes_d}
\title{Complete list of fixed discrete palettes}
\format{
A list of 66 lists.
}
\usage{
palettes_d
}
\description{
A list of all the discrete palettes of fixed lengths included in this
package. Structured as a list of lists with the first level being each
package and the second level being the palettes available in that package.
}
\details{
Contains palettes from the following packages:
ButterflyColors, DresdenColor, IslamicArt, LaCroixColoR, Manu, MapPalettes, MetBrewer, MoMAColors, NatParksPalettes, NineteenEightyR, PNWColors, Polychrome, PrettyCols, RColorBrewer, RSkittleBrewer, Redmonder, awtools, basetheme, beyonce, calecopal, colRoz, colorBlindness, colorblindr, dichromat, dutchmasters, feathers, fishualize, futurevisions, ggpomological, ggprism, ggsci, ggthemes, ggthemr, ghibli, grDevices, impressionist.colors, jcolors, khroma, lisa, miscpalettes, musculusColors, nationalparkcolors, nbapalettes, nord, ochRe, palettesForR, palettetown, pals, peRReo, rcartocolor, rockthemes, rtist, severance, soilpalettes, suffrager, tayloRswift, tidyquant, trekcolors, tvthemes, unikn, vangogh, vapeplot, vapoRwave, werpals, wesanderson, yarrr.

data.frame of palette names can be found here \code{\link{palettes_d_names}}.
}
\keyword{datasets}
