% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{palettes_d}
\alias{palettes_d}
\title{Complete list of fixed discrete palettes}
\format{A list of 23 lists.}
\usage{
palettes_d
}
\description{
A list of all the discrete palettes of fixed lengths included in this
package. Structured as a list of lists with the first level being each
package and the second level being the palettes available in that package.
}
\details{
Contains palettes from the following packages: awtools, dichromat,
dutchmasters, ggsci, ggpomological, ggthemes, ghibli, grDevices, jcolors,
LaCroixColoR, NineteenEightyR, nord, ochRe, pals, Polychrome, quickpalette,
rcartocolor, RColorBrewer, Redmonder, RSkittleBrewer, wesanderson and yarrr.

data.frame of palette names can be found here \code{\link{palettes_d_names}}.
}
\keyword{datasets}
