% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{palettes_d_names}
\alias{palettes_d_names}
\title{Data Frame of all fixed discrete palettes}
\format{
A data.frame of 2504 observations with 4
variables:
\describe{
\item{package}{character, name of package}
\item{palette}{character, name of palette}
\item{length}{integer, number of colors in palette}
\item{type}{character, type of palette}
}
}
\usage{
palettes_d_names
}
\description{
A data.frame of the names and other characteristics of the discrete palettes
of fixed lengths included in this package.
}
\details{
Contains palettes names from the following packages:
ButterflyColors, DresdenColor, IslamicArt, LaCroixColoR, Manu, MapPalettes, MetBrewer, MexBrewer, MoMAColors, NatParksPalettes, NineteenEightyR, PNWColors, Polychrome, PrettyCols, RColorBrewer, RSkittleBrewer, Rdune, Redmonder, amerika, awtools, basetheme, beyonce, blueycolors, calecopal, colRoz, colorBlindness, colorblindr, dichromat, dutchmasters, feathers, fishualize, futurevisions, ggpomological, ggprism, ggsci, ggthemes, ggthemr, ghibli, grDevices, impressionist.colors, jcolors, khroma, lisa, ltc, miscpalettes, musculusColors, nationalparkcolors, nbapalettes, nord, ochRe, palettesForR, palettetown, pals, peRReo, rcartocolor, rockthemes, rtist, severance, soilpalettes, suffrager, tayloRswift, tidyquant, trekcolors, tvthemes, unikn, vangogh, vapeplot, vapoRwave, waRhol, werpals, wesanderson, yarrr.

Full list of palette can be found here \code{\link{palettes_d}}.
}
\keyword{datasets}
