% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{palettes_d_names}
\alias{palettes_d_names}
\title{Names of all fixed discrete palettes}
\format{
A data.frame of 2027 observations with 4
variables:
\describe{
\item{package}{character, name of package}
\item{palette}{character, name of palette}
\item{length}{integer, number of colors in palette}
\item{type}{character, type of palette}
}
}
\usage{
palettes_d_names
}
\description{
A data.frame of the names and other characteristics of the discrete palettes
of fixed lengths included in this package.
}
\details{
Contains palettes names from the following packages:
awtools, basetheme, beyonce, calecopal, colorBlindness, colorblindr, colRoz, dichromat, DresdenColor, dutchmasters, fishualize, futurevisions, ggpomological, ggprism, ggsci, ggthemes, ggthemr, ghibli, grDevices, IslamicArt, jcolors, khroma, LaCroixColoR, lisa, Manu, MapPalettes, miscpalettes, nationalparkcolors, nbapalettes, NineteenEightyR, nord, ochRe, palettesForR, palettetown, pals, PNWColors, Polychrome, rcartocolor, RColorBrewer, Redmonder, rockthemes, RSkittleBrewer, rtist, soilpalettes, suffrager, tidyquant, trekcolors, tvthemes, unikn, vapeplot, vapoRwave, werpals, wesanderson, yarrr.

Full list of palette can be found here \code{\link{palettes_d}}.
}
\keyword{datasets}
